+++ PointArray.cpp
@@ -26,7 +26,7 @@ public:
 	}
 	
 	friend ostream& operator<<(ostream &TheOstream, PointArray &me) {
-		for(unsigned int i = 0; i < me.size(); i++) {
+		for(size_t i = 0; i < me.size(); i++) {
 			TheOstream << me.at(i);
 			(i < me.size() - 1) ? TheOstream << ", " : TheOstream << "";
 			TheOstream << endl;
